+++ core/src/main/java/com/graphhopper/routing/util/Bike2WeightFlagEncoder.java
@@ -50,7 +50,7 @@ public class Bike2WeightFlagEncoder extends BikeFlagEncoder {
     public void applyWayTags(ReaderWay way, EdgeIteratorState edge) {
         PointList pl = edge.fetchWayGeometry(FetchMode.ALL);
         if (!pl.is3D())
-            throw new IllegalStateException(toString() + " requires elevation data to improve speed calculation based on it. Please enable it in config via e.g. graph.elevation.provider: srtm");
+            throw new IllegalStateException(getName() + " requires elevation data to improve speed calculation based on it. Please enable it in config via e.g. graph.elevation.provider: srtm");
 
         IntsRef intsRef = edge.getFlags();
         if (way.hasTag("tunnel", "yes") || way.hasTag("bridge", "yes") || way.hasTag("highway", "steps"))
@@ -110,7 +110,7 @@ public class Bike2WeightFlagEncoder extends BikeFlagEncoder {
     }
 
     @Override
-    public String toString() {
+    public String getName() {
         return "bike2";
     }
 }