+++ core/src/main/java/com/graphhopper/routing/util/MountainBikeFlagEncoder.java
@@ -168,7 +168,7 @@ public class MountainBikeFlagEncoder extends BikeCommonFlagEncoder {
     }
 
     @Override
-    public String toString() {
+    public String getName() {
         return "mtb";
     }
 }